+++ src/sim/integration/runge_kutta.rs
@@ -7,20 +7,20 @@ use nalgebra::{matrix, vector, SMatrix, SVector};
 use super::Integrator;
 
 pub struct RungeKutta<const D: usize> {
-    a: SMatrix<f32, D, D>,
-    b: SVector<f32, D>,
-    c: SVector<f32, D>,
+    a: SMatrix<f64, D, D>,
+    b: SVector<f64, D>,
+    c: SVector<f64, D>,
 }
 
 impl<const D: usize> RungeKutta<D> {
-    pub const fn new(a: SMatrix<f32, D, D>, b: SVector<f32, D>, c: SVector<f32, D>) -> Self {
+    pub const fn new(a: SMatrix<f64, D, D>, b: SVector<f64, D>, c: SVector<f64, D>) -> Self {
         Self { a, b, c }
     }
 }
 
 impl<const D: usize> Integrator for RungeKutta<D> {
-    fn step<const R: usize>(&self, system: &mut impl crate::sim::System<R>, stepsize: f32) {
-        let mut k = SMatrix::<f32, R, D>::zeros();
+    fn step<const R: usize>(&self, system: &mut impl crate::sim::System<R>, stepsize: f64) {
+        let mut k = SMatrix::<f64, R, D>::zeros();
 
         for i in 0..D {
             // See [1] p. VI-12
@@ -31,7 +31,7 @@ impl<const D: usize> Integrator for RungeKutta<D> {
                     &(system.get_state()
                         + (0..D)
                             .map(|j| self.a[(i, j)] * k.column(j))
-                            .sum::<SVector<f32, R>>()),
+                            .sum::<SVector<f64, R>>()),
                 );
             k.set_column(i, &ki);
         }
@@ -42,7 +42,7 @@ impl<const D: usize> Integrator for RungeKutta<D> {
         let new_state = system.get_state().clone()
             + (0..D)
                 .map(|i| self.b[i] * k.column(i))
-                .sum::<SVector<f32, R>>();
+                .sum::<SVector<f64, R>>();
 
         system.set_state(new_state);
         system.set_time(system.get_time() + stepsize);
@@ -65,9 +65,9 @@ mod tests {
     use nalgebra::Vector2;
 
     pub struct Example {
-        time: f32,
+        time: f64,
         // state = [position, velocity]
-        state: Vector2<f32>,
+        state: Vector2<f64>,
     }
 
     impl Example {
@@ -78,7 +78,7 @@ mod tests {
             };
         }
 
-        fn solution(&self) -> Vector2<f32> {
+        fn solution(&self) -> Vector2<f64> {
             // x = 1/3*t^3 + t^2 + t - 0.5e^t
             // y = t^2 + 2t + 1 - 0.5e^t
             return vector![
@@ -89,22 +89,22 @@ mod tests {
     }
 
     impl System<2> for Example {
-        fn get_time(&self) -> f32 {
+        fn get_time(&self) -> f64 {
             return self.time;
         }
 
-        fn get_state(&self) -> Vector2<f32> {
+        fn get_state(&self) -> Vector2<f64> {
             return self.state;
         }
-        fn set_state(&mut self, state: Vector2<f32>) {
+        fn set_state(&mut self, state: Vector2<f64>) {
             self.state = state;
         }
 
-        fn set_time(&mut self, time: f32) {
+        fn set_time(&mut self, time: f64) {
             self.time = time;
         }
 
-        fn system(&self, time: f32, state: &Vector2<f32>) -> Vector2<f32> {
+        fn system(&self, time: f64, state: &Vector2<f64>) -> Vector2<f64> {
             // x' = y
             // y' = y - t^2 + 1
             return vector![state.y, (state.y - time.powi(2) + 1.)];
