+++ src/sim/vehicle.rs
@@ -1,18 +1,18 @@
 // Created by Tibor Völcker (tiborvoelcker@hotmail.de) on 22.11.23
-// Last modified by Tibor Völcker on 27.11.23
+// Last modified by Tibor Völcker on 28.11.23
 // Copyright (c) 2023 Tibor Völcker (tiborvoelcker@hotmail.de)
 
 use nalgebra::{vector, Vector3};
 
 pub struct Vehicle {
-    pub position: Vector3<f32>,
-    pub velocity: Vector3<f32>,
-    mass: f32,
+    pub position: Vector3<f64>,
+    pub velocity: Vector3<f64>,
+    mass: f64,
     engines: Vec<Engine>,
 }
 
 impl Vehicle {
-    pub fn new(mass: f32, engines: Vec<Engine>) -> Self {
+    pub fn new(mass: f64, engines: Vec<Engine>) -> Self {
         Self {
             position: vector![0., 0., 0.],
             velocity: vector![0., 0., 0.],
@@ -21,26 +21,26 @@ impl Vehicle {
         }
     }
 
-    pub fn thrust(&self, pressure_atmos: f32) -> Vector3<f32> {
+    pub fn thrust(&self, pressure_atmos: f64) -> Vector3<f64> {
         return self
             .engines
             .iter()
             .map(|eng| eng.thrust(pressure_atmos))
-            .sum::<Vector3<f32>>()
+            .sum::<Vector3<f64>>()
             / self.mass;
     }
 }
 
 pub struct Engine {
     // [pitch, yaw]
-    incidence: [f32; 2],
-    throttle: f32,
-    thrust_vac: f32,
-    exit_area: f32,
+    incidence: [f64; 2],
+    throttle: f64,
+    thrust_vac: f64,
+    exit_area: f64,
 }
 
 impl Engine {
-    pub fn new(incidence: [f32; 2], thrust_vac: f32, exit_area: f32) -> Self {
+    pub fn new(incidence: [f64; 2], thrust_vac: f64, exit_area: f64) -> Self {
         return Self {
             incidence,
             throttle: 1.,
@@ -49,7 +49,7 @@ impl Engine {
         };
     }
 
-    fn thrust(&self, pressure_atmos: f32) -> Vector3<f32> {
+    fn thrust(&self, pressure_atmos: f64) -> Vector3<f64> {
         return vector![
             self.incidence[1].cos() * self.incidence[0].cos(),
             self.incidence[1].sin(),
